refactor #65 : LevelService 리팩토링

+++ src/main/java/dopamine/backend/domain/level/service/LevelService.java
@@ -23,13 +23,11 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-@Slf4j
 public class LevelService {
 
     private final LevelRepository levelRepository;
     private final LevelMapper levelMapper;
     private final MemberRepository memberRepository;
-    private final MemberService memberService;
 
     public static int INF = 1000000000;
 
@@ -40,7 +38,10 @@ public class LevelService {
      */
     public Level createLevel(LevelRequestDto levelRequestDto) {
 
+        // levelNum : 하나씩 증가하도록
         int levelNum = createLevelNum();
+
+        // 이전 레벨보다 작거나, 다음 레벨보다 큰지 검증
         verifiedExp(levelNum, levelRequestDto.getExp());
 
         // create
@@ -62,13 +63,20 @@ public class LevelService {
      * @param levelId
      */
     public void deleteLevel(Long levelId) {
+
+        // 레벨 찾기
         Level level = verifiedLevel(levelId);
-        List<Member> members = memberRepository.findAllByLevel(level);
-        for(Member member : members) {
-            member.changeMember(null, null, null, 0, findLevelByLevelNum(1));
-        }
+        Level levelOne = findLevelByLevelNum(1);
+
+        // 1레벨 삭제 불가(모든 회원의 default)
+        if (level == levelOne) throw new BusinessLogicException(ExceptionCode.LEVEL_1_CANNOT_DELETE);
+
+        // 레벨에 포함된 회원을 1레벨로 변경 후 삭제
+        level.getMembers().forEach(member -> member.setLevel(levelOne));
         levelRepository.delete(level);
-        orderLevelNum();
+
+        // levelNum 정렬 & 관련 멤버의 레벨 변경
+        sortLevelNum();
         allMemberLevelChange();
     }
 
@@ -83,6 +91,27 @@ public class LevelService {
         Level level = verifiedLevel(levelId);
         LevelResponseDto levelResponseDto = levelMapper.levelToLevelResponseDto(level);
         return levelResponseDto;
+    }
+
+    /**
+     * GET : 레벨 디테일하게 조회
+     * @param member
+     * @return
+     */
+    @Transactional(readOnly = true)
+    public LevelDetailResponseDto getMemberDetailLevel(Member member) {
+        int expRange = levelRepository.findLevelByLevelNum(member.getLevel().getLevelNum() + 1).map(Level::getExp).orElse(INF) - member.getLevel().getExp();
+        int expMember = levelRepository.findLevelByLevelNum(member.getLevel().getLevelNum() + 1).map(Level::getExp).orElse(INF) - member.getExp() - 1;
+        int expPercent = (int)Math.round((double) expMember/expRange * 100);
+
+        return LevelDetailResponseDto.builder()
+                .levelId(member.getLevel().getLevelId())
+                .levelNum(member.getLevel().getLevelNum())
+                .name(member.getLevel().getName())
+                .badge(member.getLevel().getBadge())
+                .expRange(expRange)
+                .expMember(expMember)
+                .expPercent(expPercent).build();
     }
 
     /**
@@ -107,15 +136,22 @@ public class LevelService {
     }
 
     /**
-     * 검증 -> levelId 입력하면 관련 ChallengeMember Entity가 있는지 확인
+     * 검증 -> levelId 입력하면 관련 Level Entity가 있는지 확인
      * @param levelId
      * @return level
      */
+    @Transactional(readOnly = true)
     public Level verifiedLevel(Long levelId) {
         Optional<Level> level = levelRepository.findById(levelId);
         return level.orElseThrow(() -> new BusinessLogicException(ExceptionCode.LEVEL_NOT_FOUND));
     }
 
+    /**
+     * levelNum을 기준으로 Level Entity 찾기
+     * @param levelNum
+     * @return Level
+     */
+    @Transactional(readOnly = true)
     public Level findLevelByLevelNum(int levelNum) {
         Optional<Level> level = levelRepository.findLevelByLevelNum(levelNum);
         return level.orElseThrow(() -> new BusinessLogicException(ExceptionCode.LEVEL_NOT_FOUND));
@@ -124,6 +160,7 @@ public class LevelService {
     /**
      * DB에서 가장 큰 levelNum 조회 + 1 (순차적으로)
      */
+    @Transactional(readOnly = true)
     public int createLevelNum(){
         return levelRepository.findTopByOrderByLevelNumDesc()
                 .map(level -> level.getLevelNum() + 1)
@@ -135,6 +172,7 @@ public class LevelService {
      * @param levelNum
      * @param exp
      */
+    @Transactional(readOnly = true)
     public void verifiedExp(int levelNum, int exp) {
         // 이전 기준 경험치보다 큰지 검증
         int preExp = levelRepository.findLevelByLevelNum(levelNum - 1).map(Level::getExp).orElse(-1);
@@ -152,35 +190,28 @@ public class LevelService {
     /**
      * levelNum연속으로 이어지도록 설정(삭제시 levelNum한칸씩 앞으로)
      */
-    public void orderLevelNum() {
-        List<Level> levels = levelRepository.findAllByOrderByLevelNumAsc();
-        for (int i=0; i < levels.size() ; i++) {
-            levels.get(i).changeLevel(i+1, null, null, 0);
+    private void sortLevelNum() {
+        int levelNum = 0;
+        for (Level level : levelRepository.findAllByOrderByLevelNumAsc()) {
+            level.changeLevelNum(++levelNum);
         }
     }
 
-    public LevelDetailResponseDto memberDetailLevel(Member member) {
-        int expRange = levelRepository.findLevelByLevelNum(member.getLevel().getLevelNum() + 1).map(Level::getExp).orElse(INF) - member.getLevel().getExp();
-        int expMember = levelRepository.findLevelByLevelNum(member.getLevel().getLevelNum() + 1).map(Level::getExp).orElse(INF) - member.getExp() - 1;
-        int expPercent = (int)Math.round((double) expMember/expRange * 100);
-
-        return LevelDetailResponseDto.builder()
-                .levelId(member.getLevel().getLevelId())
-                .levelNum(member.getLevel().getLevelNum())
-                .name(member.getLevel().getName())
-                .badge(member.getLevel().getBadge())
-                .expRange(expRange)
-                .expMember(expMember)
-                .expPercent(expPercent).build();
-    }
-
-    // Level 변경시 모든 멤버의 레벨에 반영
+    /**
+     * Level 변경시 모든 멤버의 레벨에 반영
+     */
     private void allMemberLevelChange(){
-        List<Member> members = memberRepository.findAll();
-        for (Member member : members) {
-            Level level = memberService.getMemberLevel(member.getExp());
-            member.changeMember(null, null, null, 0, level);
-        }
+        memberRepository.findAll().forEach(member -> member.setLevel(getMemberLevel(member.getExp())));
+    }
+
+    /**
+     * 경험치에 속하는 Level 반환
+     * @param exp
+     * @return Level
+     */
+    @Transactional(readOnly = true)
+    public Level getMemberLevel(int exp) {
+        return levelRepository.findTopByExpLessThanEqualOrderByExpDesc(exp);
     }
 
 }
